Generalized sapling support and fixed #1753
