+++ src/main/java/forestry/farming/logic/FarmableVanillaSapling.java
@@ -33,63 +33,9 @@ import forestry.core.proxy.Proxies;
 import forestry.core.utils.GeneticsUtil;
 import forestry.plugins.ForestryPluginUids;
 
-public class FarmableVanillaSapling implements IFarmable {
-
-	protected final ItemStack germling;
-	private final ItemStack[] windfall;
+public class FarmableVanillaSapling extends FarmableSapling {
 
 	public FarmableVanillaSapling() {
-		this.germling = new ItemStack(Blocks.SAPLING);
-		this.windfall = new ItemStack[]{new ItemStack(Items.APPLE), new ItemStack(FarmableCocoa.COCOA_SEED, 1, FarmableCocoa.COCOA_META)};
-	}
-
-	@Override
-	public boolean plantSaplingAt(EntityPlayer player, ItemStack germling, World world, BlockPos pos) {
-		if (ForestryAPI.enabledPlugins.contains(ForestryPluginUids.ARBORICULTURE)) {
-			IIndividual tree = GeneticsUtil.getGeneticEquivalent(germling);
-			if (!(tree instanceof ITree)) {
-				return false;
-			}
-
-			return TreeManager.treeRoot.plantSapling(world, (ITree) tree, player.getGameProfile(), pos);
-		} else {
-			EnumActionResult actionResult = germling.copy().onItemUse(player, world, pos.down(), EnumHand.MAIN_HAND, EnumFacing.UP, 0, 0, 0);
-			if (actionResult == EnumActionResult.SUCCESS) {
-				PacketFXSignal packet = new PacketFXSignal(PacketFXSignal.SoundFXType.BLOCK_PLACE, pos, Blocks.SAPLING.getDefaultState());
-				Proxies.net.sendNetworkPacket(packet, world);
-				return true;
-			}
-			return false;
-		}
-	}
-
-	@Override
-	public boolean isSaplingAt(World world, BlockPos pos) {
-		return world.getBlockState(pos).getBlock() == Blocks.SAPLING;
-	}
-
-	@Override
-	public ICrop getCropAt(World world, BlockPos pos, IBlockState blockState) {
-		Block block = blockState.getBlock();
-		if (!block.isWood(world, pos)) {
-			return null;
-		}
-
-		return new CropDestroy(world, blockState, pos, null);
-	}
-
-	@Override
-	public boolean isGermling(ItemStack itemstack) {
-		return ItemStack.areItemsEqual(germling, itemstack);
-	}
-
-	@Override
-	public boolean isWindfall(ItemStack itemstack) {
-		for (ItemStack drop : windfall) {
-			if (drop.isItemEqual(itemstack)) {
-				return true;
-			}
-		}
-		return false;
+		super(new ItemStack(Blocks.SAPLING), new ItemStack[]{new ItemStack(Items.APPLE), new ItemStack(FarmableCocoa.COCOA_SEED, 1, FarmableCocoa.COCOA_META)});
 	}
 }
